requirements.txt
Flask==3.0.3
Flask_Cors==3.0.10
keras==3.4.1
numpy==1.26.4
pandas==2.2.2
psycopg2_binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
Requests==2.32.3
scipy==1.14.0
